- Fixed bug that caused (empty) coverage to be generated even for `--method=nette-tester-list-methods`, improving startup time for codebases with a lot of test cases
- <del>Reduced blocking time of coverage to 3 % of original.</del> (merged in https://github.com/nette/tester/commit/9b62462d832f00f881fb84857986bfb51519505f)

Collector benchmark:
```
original blocking: 74.48 ms

     new prepared: 71.04 ms
     new blocking:  2.35 ms
          new sum: 73.38 ms
```