@milo 2015-08-10
What about not to set `Environment::COVERAGE` in [initial](https://github.com/nette/tester/blob/fc6d85565097528168ebd9f106744a22d8155ad1/src/Runner/TestHandler.php#L143) phase?