@Mikulas 2015-08-10
I've looked into that but it seemed it would require a lot of refactoring, which I deemed unnecessary.