@Mikulas 2016-02-12
This was partially incorporated by https://github.com/nette/tester/commit/9b62462d832f00f881fb84857986bfb51519505f .

----

> bug that caused (empty) coverage to be generated even for --method=nette-tester-list-methods, improving startup time for codebases with a lot of test cases

still needs to be fixed though.