@milo 2016-05-19
@Mikulas Do you think that something from this PR is needed after bef60c6e58cecee3349d47f7736000e709c37400?