@Mikulas 2016-05-19
I believe everything from this PR was merged. Thanks :)